Moved Task stylesheets into new /styles/ subdir
